Move error handling functions to `modules/errors.ts`
Extract showGlobalErrorMessage, shouldIgnoreError, and
processWindowErrorEvent into a side-effect-free module. This keeps
bootstrap.ts minimal (just the init guard and side effect) and prevents
the ESM bundle from re-triggering the global error handler.
