+++ web_src/js/modules/errors.test.ts
@@ -1,4 +1,15 @@
-import {shouldIgnoreError} from './bootstrap.ts';
+import {showGlobalErrorMessage, shouldIgnoreError} from './errors.ts';
+
+test('showGlobalErrorMessage', () => {
+  document.body.innerHTML = '<div class="page-content"></div>';
+  showGlobalErrorMessage('test msg 1');
+  showGlobalErrorMessage('test msg 2');
+  showGlobalErrorMessage('test msg 1'); // duplicated
+
+  expect(document.body.innerHTML).toContain('>test msg 1 (2)<');
+  expect(document.body.innerHTML).toContain('>test msg 2<');
+  expect(document.querySelectorAll('.js-global-error').length).toEqual(2);
+});
 
 test('shouldIgnoreError', () => {
   for (const url of [
