--- web_src/js/modules/message.test.ts
@@ -1,12 +1,0 @@
-import {showGlobalErrorMessage} from './message.ts';
-
-test('showGlobalErrorMessage', () => {
-  document.body.innerHTML = '<div class="page-content"></div>';
-  showGlobalErrorMessage('test msg 1');
-  showGlobalErrorMessage('test msg 2');
-  showGlobalErrorMessage('test msg 1'); // duplicated
-
-  expect(document.body.innerHTML).toContain('>test msg 1 (2)<');
-  expect(document.body.innerHTML).toContain('>test msg 2<');
-  expect(document.querySelectorAll('.js-global-error').length).toEqual(2);
-});
